+++ pkgs/xsoulspace_review/lib/src/store_reviewer/reviewers/reviewers.dart
@@ -1,5 +1,0 @@
-export 'google_apple_store_reviewer.dart';
-export 'huawei_sotre_reviewer.dart';
-export 'rustore_reviewer.dart';
-export 'snapstore_reviewer.dart';
-export 'web_store_reviewer.dart';