+++ pkgs/xsoulspace_review_google_apple/lib/xsoulspace_review_google_apple.dart
@@ -1,0 +1,2 @@
+export 'src/google_apple_store_reviewer.dart';
+