+++ pkgs/xsoulspace_review_huawei/lib/xsoulspace_review_huawei.dart
@@ -1,0 +1,2 @@
+export 'src/huawei_store_reviewer.dart';
+