+++ pkgs/xsoulspace_review_rustore/lib/xsoulspace_review_rustore.dart
@@ -1,0 +1,2 @@
+export 'src/rustore_reviewer.dart';
+