+++ pkgs/xsoulspace_review_snapstore/lib/src/snapstore_reviewer.dart
@@ -1,13 +1,21 @@
 import 'package:flutter/widgets.dart';
+import 'package:xsoulspace_review_interface/xsoulspace_review_interface.dart';
 
-import '../store_reviewer.dart';
-
+/// {@template snapstore_reviewer}
+/// Store reviewer implementation for Linux Snap Store.
+///
+/// Launches the snap review scheme with a consent dialog.
+/// {@endtemplate}
 final class SnapStoreReviewer extends StoreReviewer {
-  SnapStoreReviewer({
+  /// {@macro snapstore_reviewer}
+  const SnapStoreReviewer({
     required super.packageName,
-    super.consentBuilder,
+    required this.consentBuilder,
     super.defaultLocale,
   });
+
+  /// A builder for the consent screen before opening Snap Store
+  final ReviewerFallbackConsentBuilder consentBuilder;
 
   @override
   Future<bool> onLoad() async => true;