+++ pkgs/xsoulspace_review_snapstore/lib/xsoulspace_review_snapstore.dart
@@ -1,0 +1,2 @@
+export 'src/snapstore_reviewer.dart';
+