+++ pkgs/xsoulspace_review_web/lib/xsoulspace_review_web.dart
@@ -1,0 +1,2 @@
+export 'src/web_store_reviewer.dart';
+